add different duration for each test
