+++ packages/k6-tests/src/values/env.md
@@ -361,6 +361,17 @@ specifies the condition for failed HTTP requests, only necessary if `ENABLE_THRE
 ## ENV `THRESHOLD_HTTP_REQ_DURATION`
 
 specifies the condition for the duration of HTTP requests, only necessary if `ENABLE_THRESHOLDS` is set to `true`.
+The duration can be different for each test, as there are tests where we download files for example, so `http_req_duration` is set as a separate environment variable for each test.
+ 
+example: 
+
+```
+thresholds: {
+    http_req_failed: [settings.thresholds.rate],
+    http_req_duration: ENV('TEST_KOKO_PLATFORM_080_RAMPING_THRESHOLDS_DURATION', 'p(95)<300'),
+},
+```
+
 
 * default value: `p(95)<200` (indicating that 95% of requests should be completed in less than 200ms)
 * `export THRESHOLD_HTTP_REQ_DURATION='p(95)<200'`
